+++ HISTORY.md
@@ -1,10 +1,18 @@
 ## Change Log
 
 
-#### [v0.0.2](https://github.com/aplazame/woocommerce/tree/v0.0.2) (2015-?-?)
+#### [v0.0.3](https://github.com/aplazame/woocommerce/tree/v0.0.3) (2015-??-??)
 
-[Full Changelog v0.0.2-dev](https://github.com/aplazame/woocommerce/compare/v0.0.1...v0.0.2)
+[Full Changelog v0.0.3](https://github.com/aplazame/woocommerce/compare/v0.0.1...v0.0.3)
 
+* fix ?
+
+
+#### [v0.0.2](https://github.com/aplazame/woocommerce/tree/v0.0.2) (2015-09-14)
+
+[Full Changelog v0.0.2](https://github.com/aplazame/woocommerce/compare/v0.0.1...v0.0.2)
+
+* l10n fix and review
 * API exception handler
 * process_refund on gateway class
 * Full Readme.txt